refactor(app): extract NddScoreGeneDetail data layer into composable (#397)
Move the gene-detail resource loading, JSON parsing, and derived computed state
out of NddScoreGeneDetail.vue (750 -> 550 lines) into useNddScoreGeneDetail.ts.
The component keeps every disclosure/help copy string verbatim ('ML prediction',
'Machine learning, not manual curation', 'Not a curated SysNDD gene', etc.) and
the same props/render output.

Advances #397. Part of #346.
